support deleting multiple databases with turso db destroy fixes https://github.com/tursodatabase/turso-cli/issues/667

+++ internal/cmd/db_destroy.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"fmt"
+	"strings"
 
 	"github.com/chiselstrike/iku-turso-cli/internal"
 	"github.com/spf13/cobra"
@@ -18,7 +19,7 @@ func init() {
 var destroyCmd = &cobra.Command{
 	Use:               "destroy database_name",
 	Short:             "Destroy a database.",
-	Args:              cobra.ExactArgs(1),
+	Args:              cobra.MinimumNArgs(1),
 	ValidArgsFunction: dbNameArg,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		cmd.SilenceUsage = true
@@ -27,31 +28,44 @@ var destroyCmd = &cobra.Command{
 			return err
 		}
 
-		db, err := getDatabase(client, args[0])
-		if err != nil {
-			return nil
-		}
+		if instanceFlag != "" || locationFlag != "" {
+			for _, name := range args {
+				db, err := getDatabase(client, args[0])
+				if err != nil {
+					return nil
+				}
 
-		name := args[0]
-		if instanceFlag != "" {
-			if db.Group != "" {
-				return fmt.Errorf("group databases do not support instance destruction.\nUse %s instead", internal.Emph("turso group locations rm "+name))
-			}
-			return destroyDatabaseInstance(client, name, instanceFlag)
-		}
+				if instanceFlag != "" {
+					if db.Group != "" {
+						return fmt.Errorf("group databases do not support instance destruction.\nUse %s instead", internal.Emph("turso group locations rm "+name))
+					}
 
-		if locationFlag != "" {
-			if db.Group != "" {
-				return fmt.Errorf("group databases do not support location destruction.\nUse %s instead", internal.Emph("turso group locations rm "+name+" "+locationFlag))
+					if err := destroyDatabaseInstance(client, name, instanceFlag); err != nil {
+						return err
+					}
+				}
+
+				if locationFlag != "" {
+					if db.Group != "" {
+						return fmt.Errorf("group databases do not support location destruction.\nUse %s instead", internal.Emph("turso group locations rm "+name+" "+locationFlag))
+					}
+
+					if err := destroyDatabaseRegion(client, name, locationFlag); err != nil {
+						return err
+					}
+				}
 			}
-			return destroyDatabaseRegion(client, name, locationFlag)
 		}
 
 		if yesFlag {
-			return destroyDatabase(client, name)
+			return destroyDatabases(client, args)
 		}
 
-		fmt.Printf("Database %s, and all its data will be destroyed.\n", internal.Emph(name))
+		if len(args) > 0 {
+			fmt.Printf("Databases %s and all their data will be destroyed.\n", internal.Emph(strings.Join(args, ", ")))
+		} else {
+			fmt.Printf("Database %s, and all its data will be destroyed.\n", internal.Emph(args[0]))
+		}
 
 		ok, err := promptConfirmation("Are you sure you want to do this?")
 		if err != nil {
@@ -63,6 +77,6 @@ var destroyCmd = &cobra.Command{
 			return nil
 		}
 
-		return destroyDatabase(client, name)
+		return destroyDatabases(client, args)
 	},
 }
